mvvm+stateFlow+viewBinding+Flow+Coroutines
