+++ app/src/main/java/com/hfad/currencycalculator/data/remote/RetrofitCurrencyListDataSource.kt
@@ -1,0 +1,10 @@
+package com.hfad.currencycalculator.data.remote
+
+import com.hfad.currencycalculator.domain.repository.CurrencyListRemoteDataSource
+import com.hfad.currencycalculator.data.remote.models.RemoteValuteResponce
+import com.hfad.currencycalculator.data.network.CurrencyApi
+
+class RetrofitCurrencyListDataSource(private val mApi: CurrencyApi) : CurrencyListRemoteDataSource {
+
+    override suspend fun getCurrencyList(): RemoteValuteResponce = mApi.fetchValute()
+}
