+++ app/src/main/java/com/hfad/currencycalculator/di/NetModule.kt
@@ -1,0 +1,71 @@
+package com.hfad.currencycalculator.di
+
+import android.util.Log
+import com.google.gson.Gson
+
+import com.hfad.currencycalculator.data.network.CurrencyApi
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+
+import retrofit2.Retrofit
+import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
+import retrofit2.converter.gson.GsonConverterFactory
+
+import javax.inject.Singleton
+
+
+@Module
+@InstallIn(SingletonComponent::class)
+class NetModule() {
+
+    @Provides
+    @Singleton
+    fun httpLoggingInterceptor(): HttpLoggingInterceptor =
+        HttpLoggingInterceptor().apply {
+            this.level = HttpLoggingInterceptor.Level.BODY
+        }
+
+
+    @Provides
+    @Singleton
+    fun okHttpClient(
+        httpLoggingInterceptor: HttpLoggingInterceptor,
+    ): OkHttpClient =
+        OkHttpClient.Builder()
+            .addInterceptor(httpLoggingInterceptor)
+            .build()
+
+    @Provides
+    @Singleton
+    fun provideGson(): Gson = Gson()
+
+    @Provides
+    @Singleton
+    fun provideRetrofit(gson: Gson, client: OkHttpClient): Retrofit =
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .client(client)
+            .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
+            .addConverterFactory(GsonConverterFactory.create(gson))
+            .build()
+
+    @Provides
+    @Singleton
+    fun provideApi(retrofit: Retrofit): CurrencyApi =
+        retrofit.create(CurrencyApi::class.java)
+
+//    @Provides
+//    @Singleton
+//    fun providePresenter(mCurrencyListRepository: CurrencyListRepository): CurrencyPresenter
+//        {
+//            Log.e("SDSD", "providePresenter: ${CurrencyPresenter(mCurrencyListRepository)}", )
+//            return CurrencyPresenter(mCurrencyListRepository)
+//        }
+}
+
+private const val BASE_URL = "https://www.cbr-xml-daily.ru/"
