+++ app/src/main/java/com/hfad/currencycalculator/domain/repository/CurrencyListRemoteDataSource.kt
@@ -1,0 +1,7 @@
+package com.hfad.currencycalculator.domain.repository
+
+import com.hfad.currencycalculator.data.remote.models.RemoteValuteResponce
+
+interface CurrencyListRemoteDataSource {
+    suspend fun getCurrencyList(): RemoteValuteResponce
+}
