--- app/src/main/java/com/hfad/currencycalculator/model/local/CurrencyListRemoteDataSource.kt
@@ -1,8 +1,0 @@
-package com.hfad.currencycalculator.model.local
-
-import com.hfad.currencycalculator.model.remote.RemoteValuteResponce
-import io.reactivex.Single
-
-interface CurrencyListRemoteDataSource {
-    fun getCurrencyList():Single<RemoteValuteResponce>
-}
