--- app/src/main/java/com/hfad/currencycalculator/model/remote/RemoteValuteResponce.kt
@@ -1,11 +1,0 @@
-package com.hfad.currencycalculator.model.remote
-
-import com.google.gson.annotations.SerializedName
-
-data class RemoteValuteResponce(
-    val Date: String,
-    val PreviousDate: String,
-    val PreviousURL: String,
-    val Timestamp: String,
-    val Valute: Map<String, RemoteValute>
-)
