--- app/src/main/java/com/hfad/currencycalculator/model/room/dao/CurrencyEntity.kt
@@ -1,27 +1,0 @@
-package com.hfad.currencycalculator.model.room.dao
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.hfad.currencycalculator.model.room.dao.CurrencyEntity.Companion.TABLE_NAME
-
-@Entity(tableName = TABLE_NAME)
-data class CurrencyEntity(
-    @ColumnInfo(name = "date")
-    val Date: String,
-    @ColumnInfo(name = "prev_date")
-    val PreviousDate: String,
-    @ColumnInfo(name = "prev_url")
-    val PreviousURL: String,
-    @ColumnInfo(name = "time_stamp")
-    val Timestamp: String,
-) {
-
-    @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "ID")
-    var id: Int = 0
-
-    companion object {
-        const val TABLE_NAME = "currency_entity_table"
-    }
-}
