--- app/src/main/java/com/hfad/currencycalculator/model/state/CurrencyEnum.kt
@@ -1,8 +1,0 @@
-package com.hfad.currencycalculator.model.state
-
-import androidx.room.ColumnInfo
-
-enum class CurrencyEnum{
-     LEFTCURRENCY,
-     RIGHTCURRENCY
-}
