--- app/src/main/java/com/hfad/currencycalculator/navigation/BaseNavigation.kt
@@ -1,7 +1,0 @@
-package com.hfad.currencycalculator.navigation
-
-import androidx.fragment.app.Fragment
-
-interface BaseNavigation {
-    fun openFragment(x: Fragment): Unit
-}
