+++ app/src/main/java/com/hfad/currencycalculator/presentation/currency_screen/currency_state/CurrencyEnum.kt
@@ -1,0 +1,8 @@
+package com.hfad.currencycalculator.presentation.currency_screen.currency_state
+
+import androidx.room.ColumnInfo
+
+enum class CurrencyEnum{
+     LEFTCURRENCY,
+     RIGHTCURRENCY
+}
