+++ app/src/main/java/com/hfad/currencycalculator/presentation/navigation/BaseNavigation.kt
@@ -1,0 +1,7 @@
+package com.hfad.currencycalculator.presentation.navigation
+
+import androidx.fragment.app.Fragment
+
+interface BaseNavigation {
+    fun openFragment(x: Fragment)
+}
